requirements.txt
flask
paddleocr
paddlepaddle
opencv-python
numpy<2.0
python-docx
pdf2image
sqlalchemy